Remove short flag for passphrase (#158)
We would like to keep the short flags, in the command line, only for everyday use and secure flags.

As we consider that entering the passphrase directly in the command line is not a good idea (see below), but still want to give this opportunity (test, demo and such), we would like to only have the long flag for passphrase (--password), which means, no more -p.

This is supposed to give a hint that we don't want to ease the use of a flag we consider insecure for production.

Why this short flag is not a good idea ?

If you are into scripting, having

```
vegawallet key generate -n my-wallet -p "unbreakable-pa44w0rD"
```

expose your passphrase to everyone that might gain access to you script, adding security holes.

Also, if you execute this command it into your terminal, walking the shell history will also reveal your passphrase, and might be retrieved by other programs.