Support wallet isolation
On HD wallets, the wallet node is used to generate (and retrieve) key pairs. For security purpose, you don't want to store the wallet node on the machine running the node, because it can be compromised. So, now, you can isolate a single key pair, without the wallet node, in an "isolated wallet".