fix: Remove unnecessary verification of the transaction
This verification is unecessary as we sing the input transaction not the one coming from the UI.

If we were in a situation where the third party and the UI play together to *succesfully* forge a request, that would mean the entire verification process is compromised. Thus, at this point, this verification would be easily bypassed anyway.